Separated handlers and routes into separate files; added 'coverage' to .gitignore
