+++ app/server.js
@@ -1,9 +1,10 @@
 const hapi = require('hapi');
 const pkg = require('./package');
 const R = require('ramda');
-const handleWithCatch = require('./utils/handleWithCatch');
-const catchError = require('./utils/catchError');
-const handlers = require('./lib/handlers/handlers');
+const handleWithCatch = require('./lib/utils/handleWithCatch');
+const catchError = require('./lib/utils/catchError');
+const handlers = require('./lib/handlers');
+const routes = require('./lib/routes');
 
 const server = new hapi.Server();
 
@@ -14,47 +15,13 @@ server.connection({
 
 server.route({
   method: 'GET',
-  path: '/hello',
-  handler: (request, reply) => reply('Hello, world!')
-});
-
-server.route({
-  method: 'GET',
   path: '/healthcheck',
   handler: (request, reply) => reply({ status: 'OK', version: pkg.version })
 });
 
 const safeHandlers = R.map(handleWithCatch(catchError), handlers);
 
-server.route({
-  method: 'GET',
-  path: '/task',
-  handler: safeHandlers.listHandler
-});
-
-server.route({
-  method: 'GET',
-  path: '/task/{id}',
-  handler: safeHandlers.getHandler
-});
-
-server.route({
-  method: 'POST',
-  path: '/task',
-  handler: safeHandlers.postHandler
-});
-
-server.route({
-  method: 'PUT',
-  path: '/task/{id}',
-  handler: safeHandlers.putHandler
-});
-
-server.route({
-  method: 'DELETE',
-  path: '/task/{id}',
-  handler: safeHandlers.deleteHandler
-});
+server.route(routes(safeHandlers));
 
 if (!module.parent) {
   server.start((err) => {
